+++ listar_transportadoras/models.py
@@ -8,7 +8,7 @@ class Transportadoras(models.Model):
     endereco = models.CharField(max_length=50, blank=True, null=True)
     telefone = models.CharField(max_length=50, blank=True, null=True)
     cidade = models.CharField(max_length=50, blank=True, null=True)
-    estadoid = models.IntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
+    estadoid = models.SmallIntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
     cep = models.CharField(max_length=10, blank=True, null=True)
     cnpj = models.CharField(max_length=19, blank=True, null=True)
