+++ login/models.py
@@ -9,7 +9,7 @@ class Clientes(models.Model):
     complemento = models.CharField(max_length=30, blank=True, null=True)
     numero = models.CharField(max_length=15, blank=True, null=True)
     cidade = models.CharField(max_length=50, blank=True, null=True)
-    estadoid = models.IntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
+    estadoid = models.SmallIntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
     cep = models.CharField(max_length=10, blank=True, null=True)
     ddd = models.CharField(max_length=3, blank=True, null=True)
     telefone = models.CharField(max_length=10, blank=True, null=True)
@@ -31,8 +31,8 @@ class Fornecedores(models.Model):
     nomefornecedor = models.CharField(max_length=50, blank=True, null=True)
     endereco = models.CharField(max_length=50, blank=True, null=True)
     cidade = models.CharField(max_length=50, blank=True, null=True)
-    estadoid = models.IntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
-    ddd = models.IntegerField(blank=True, null=True)
+    estadoid = models.SmallIntegerField(db_column='estadoID', blank=True, null=True)  # Field name made lowercase.
+    ddd = models.SmallIntegerField(blank=True, null=True)
     telefone = models.CharField(max_length=14, blank=True, null=True)
     usuario = models.CharField(max_length=20, blank=True, null=True)
     senha = models.CharField(max_length=20, blank=True, null=True)
